+++ app/src/main/AndroidManifest.xml
@@ -14,13 +14,16 @@
         android:theme="@style/Theme.Material3.Dark"
         tools:targetApi="31">
         <activity
-            android:name=".Register"
-            android:theme="@style/Theme.Material3.Dark.NoActionBar"
+            android:name=".Notifications"
             android:exported="false" />
         <activity
+            android:name=".Register"
+            android:exported="false"
+            android:theme="@style/Theme.Material3.Dark.NoActionBar" />
+        <activity
             android:name=".Login"
-            android:theme="@style/Theme.Material3.Dark.NoActionBar"
-            android:exported="true">
+            android:exported="true"
+            android:theme="@style/Theme.Material3.Dark.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -45,9 +48,7 @@
         <activity
             android:name=".DashboardActivity"
             android:exported="true"
-            android:label="CS4084-Project">
-
-        </activity>
+            android:label="CS4084-Project"></activity>
     </application>
 
 </manifest>
