+++ app/src/main/res/layout/fragment_notifications.xml
@@ -9,7 +9,7 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Post Fragment"
+        android:text="Notifications Fragment"
         android:textSize="25dp"
         android:layout_centerInParent="true"  />
