+++ app/src/main/res/menu/menu_nav.xml
@@ -5,13 +5,18 @@
         android:icon="@drawable/ic_home"
         android:title="Home" />
     <item
-        android:id="@+id/nav_post"
+        android:id="@+id/notifications"
         android:icon="@drawable/ic_post"
-        android:title="Post" />
+        android:title="Notifications" />
     <item
         android:id="@+id/nav_profile"
         android:icon="@drawable/ic_account"
         android:title="Profile" />
+
+    <item
+        android:id="@+id/post"
+        android:icon="@drawable/ic_post"
+        android:title="Post"/>
 
     <item
         android:id="@+id/logout"
