+++ app/src/main/res/values/strings.xml
@@ -15,5 +15,7 @@
     <string name="register">Register</string>
     <string name="click_to_login">Click To Login</string>
     <string name="click_to_register">Click To Register</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 
 </resources>
